fix: various fixes and tests (#1807)
* fix: module init

* fix: Various small production fixes

* test: minor fixes

* chore: publish 4.0.0-alpha.307

* feat: wip css

* feat: various fixes and tests

* feat: fixes and tests

* fix: making typechecker happy

* feat: electron launcher

* fix: electron fixes

* docs: electron changelog

* docs: changelog

+++ src/devServer/server.ts
@@ -1,41 +1,61 @@
-import { ChildProcess, fork } from 'child_process';
-
+import { ChildProcess, spawn } from 'child_process';
 import { BundleType, IBundleWriteResponse } from '../bundle/bundle';
 import { Context } from '../core/context';
+import { onExit } from '../utils/exit';
 
 export interface IServerProcess {
-  start: (cliArgs: Array<string>) => ChildProcess;
+  start: (cliArgs?: Array<string>) => ChildProcess;
   stop: () => void;
 }
 
-export const createServerProcess = (ctx: Context, bundles: Array<IBundleWriteResponse>): IServerProcess => {
-  let server: ChildProcess;
+function getAbsPath(ctx: Context, bundles: Array<IBundleWriteResponse>) {
   let entry: string;
-  if (bundles.length === 1) {
-    entry = bundles[0].absPath;
-  } else {
-    let resolved = false;
-    for (const bundle of bundles) {
-      if (bundle.bundle.type === BundleType.JS_SERVER_ENTRY) {
+  let resolved = false;
+  for (const bundle of bundles) {
+    if (bundle.bundle.type === BundleType.JS_SERVER_ENTRY) {
+      resolved = true;
+      entry = bundle.absPath;
+      break;
+    } else {
+      if (bundle.bundle.type === BundleType.JS_APP) {
         resolved = true;
         entry = bundle.absPath;
-        break;
       }
-    }
-    if (!resolved) {
-      throw new Error('No serverEntry file found!!');
     }
   }
 
-  return {
-    start: (cliArgs: Array<string> = []): ChildProcess => {
-      server = fork(entry, cliArgs, {
+  if (!resolved) {
+    ctx.fatal('server', ['Failed to resolve server entry']);
+  }
+
+  return entry;
+}
+
+export const createServerProcess = (props: {
+  bundles: Array<IBundleWriteResponse>;
+  ctx: Context;
+  processName: string;
+}): IServerProcess => {
+  const { bundles, ctx, processName } = props;
+  let server: ChildProcess;
+  if (ctx.serverProcess) return ctx.serverProcess;
+
+  let entry = getAbsPath(ctx, bundles);
+
+  //const electronPath = require('electron');
+
+  const self = {
+    start: (cliArgs?: Array<string>): ChildProcess => {
+      cliArgs = cliArgs || [];
+      ctx.log.info('server', `spawn ${entry}`);
+      server = spawn(processName, [entry, ...cliArgs], {
         env: {
           ...process.env,
           ...ctx.config.env,
         },
         stdio: 'inherit',
       });
+
       server.on('close', code => {
         if (code === 8) {
           console.error('Error detected, waiting for changes...');
@@ -44,7 +64,23 @@ export const createServerProcess = (ctx: Context, bundles: Array<IBundleWriteRes
       return server;
     },
     stop: () => {
-      server.kill();
+      if (server) {
+        server.kill();
+        ctx.log.info('server', `Killed ${entry}`);
+      }
     },
   };
+  onExit('ServerProcess', () => {
+    self.stop();
+  });
+
+  ctx.ict.on('rebundle', props => {
+    entry = getAbsPath(ctx, props.bundles);
+    self.stop();
+    self.start();
+  });
+
+  ctx.serverProcess = self;
+
+  return self;
 };
